+++ ChangeLog.md
@@ -21,6 +21,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - deprecated `AbstractArangoConfiguration`
 
+### Removed
+
+- removed use of javax.xml.bind to better support Java 9 and above
+
 ### Fixed
 
 - fixed NPE when loading non-existing data into fields of primitive type (issue #127)